break up lex into modules
