+++ src/lex/cpp.rs
@@ -1,0 +1,7 @@
+use super::{Lexer, Token};
+
+impl Lexer<'_> {
+    pub(super) fn cpp(&mut self) -> Token {
+        unimplemented!("preprocessing");
+    }
+}
